+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/ImpTest.hs
@@ -375,6 +375,7 @@ unRegisterDRep drep = do
 genUnRegTxCert ::
   forall era.
   ( ShelleyEraImp era
+  , ShelleyEraTxCert era
   , ConwayEraTxCert era
   ) =>
   Credential 'Staking ->
@@ -392,6 +393,7 @@ genUnRegTxCert stakingCredential = do
 genRegTxCert ::
   forall era.
   ( ShelleyEraImp era
+  , ShelleyEraTxCert era
   , ConwayEraTxCert era
   ) =>
   Credential 'Staking ->
@@ -511,7 +513,10 @@ setupPoolWithStake delegCoin = do
   pure (khPool, credDelegatorPayment, credDelegatorStaking)
 
 setupPoolWithoutStake ::
-  (ShelleyEraImp era, ConwayEraTxCert era) =>
+  ( ShelleyEraImp era
+  , ShelleyEraTxCert era
+  , ConwayEraTxCert era
+  ) =>
   ImpTestM era (KeyHash 'StakePool, Credential 'Staking)
 setupPoolWithoutStake = do
   khPool <- freshKeyHash
@@ -686,6 +691,7 @@ submitFailingProposal proposal expectedFailure =
 -- multiple actions in the same transaciton use `trySubmitGovActions` instead.
 trySubmitGovAction ::
   ( ShelleyEraImp era
+  , ShelleyEraTxCert era
   , ConwayEraTxBody era
   ) =>
   GovAction era ->
@@ -717,7 +723,10 @@ submitAndExpireProposalToMakeReward stakingC = do
 
 -- | Submits a transaction that proposes the given governance action
 trySubmitGovActions ::
-  (ShelleyEraImp era, ConwayEraTxBody era) =>
+  ( ShelleyEraImp era
+  , ShelleyEraTxCert era
+  , ConwayEraTxBody era
+  ) =>
   NE.NonEmpty (GovAction era) ->
   ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era)), Tx era) (Tx era))
 trySubmitGovActions gas = do
@@ -741,7 +750,10 @@ mkProposalWithRewardAccount ga rewardAccount = do
       }
 
 mkProposal ::
-  (ShelleyEraImp era, ConwayEraTxBody era) =>
+  ( ShelleyEraImp era
+  , ShelleyEraTxCert era
+  , ConwayEraTxBody era
+  ) =>
   GovAction era ->
   ImpTestM era (ProposalProcedure era)
 mkProposal ga = do
@@ -751,6 +763,7 @@ mkProposal ga = do
 submitGovAction ::
   forall era.
   ( ShelleyEraImp era
+  , ShelleyEraTxCert era
   , ConwayEraTxBody era
   , HasCallStack
   ) =>
@@ -763,6 +776,7 @@ submitGovAction ga = do
 submitGovAction_ ::
   forall era.
   ( ShelleyEraImp era
+  , ShelleyEraTxCert era
   , ConwayEraTxBody era
   , HasCallStack
   ) =>
@@ -773,6 +787,7 @@ submitGovAction_ = void . submitGovAction
 submitGovActions ::
   forall era.
   ( ShelleyEraImp era
+  , ShelleyEraTxCert era
   , ConwayEraTxBody era
   , HasCallStack
   ) =>
@@ -792,6 +807,7 @@ mkTreasuryWithdrawalsGovAction wdrls =
 
 submitTreasuryWithdrawals ::
   ( ShelleyEraImp era
+  , ShelleyEraTxCert era
   , ConwayEraTxBody era
   , ConwayEraGov era
   ) =>
@@ -801,7 +817,9 @@ submitTreasuryWithdrawals wdrls =
   mkTreasuryWithdrawalsGovAction wdrls >>= submitGovAction
 
 enactTreasuryWithdrawals ::
-  ConwayEraImp era =>
+  ( ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   [(RewardAccount, Coin)] ->
   Credential 'DRepRole ->
   NonEmpty (Credential 'HotCommitteeRole) ->
@@ -814,7 +832,9 @@ enactTreasuryWithdrawals withdrawals dRep cms = do
   pure gaId
 
 submitParameterChange ::
-  ConwayEraImp era =>
+  ( ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   StrictMaybe GovActionId ->
   PParamsUpdate era ->
   ImpTestM era GovActionId
@@ -845,6 +865,7 @@ getGovPolicy =
 submitFailingGovAction ::
   forall era.
   ( ShelleyEraImp era
+  , ShelleyEraTxCert era
   , ConwayEraTxBody era
   , HasCallStack
   ) =>
@@ -1265,6 +1286,7 @@ submitCommitteeElection ::
   forall era.
   ( HasCallStack
   , ConwayEraImp era
+  , ShelleyEraTxCert era
   ) =>
   StrictMaybe (GovPurposeId 'CommitteePurpose) ->
   Credential 'DRepRole ->
@@ -1287,6 +1309,7 @@ electBasicCommittee ::
   forall era.
   ( HasCallStack
   , ConwayEraImp era
+  , ShelleyEraTxCert era
   ) =>
   ImpTestM
     era
@@ -1431,6 +1454,7 @@ submitGovActionForest submitAction p forest =
 enactConstitution ::
   forall era.
   ( ConwayEraImp era
+  , ShelleyEraTxCert era
   , HasCallStack
   ) =>
   StrictMaybe (GovPurposeId 'ConstitutionPurpose) ->
@@ -1457,7 +1481,9 @@ expectNumDormantEpochs expected = do
   nd `shouldBeExpr` expected
 
 mkConstitutionProposal ::
-  ConwayEraImp era =>
+  ( ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   StrictMaybe (GovPurposeId 'ConstitutionPurpose) ->
   ImpTestM era (ProposalProcedure era, Constitution era)
 mkConstitutionProposal prevGovId = do
@@ -1466,7 +1492,9 @@ mkConstitutionProposal prevGovId = do
 
 submitConstitution ::
   forall era.
-  ConwayEraImp era =>
+  ( ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   StrictMaybe (GovPurposeId 'ConstitutionPurpose) ->
   ImpTestM era GovActionId
 submitConstitution prevGovId = do
@@ -1596,7 +1624,9 @@ submitYesVoteCCs_ committeeMembers govId =
   mapM_ (\c -> submitYesVote_ (CommitteeVoter c) govId) committeeMembers
 
 mkUpdateCommitteeProposal ::
-  ConwayEraImp era =>
+  ( ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   -- | Set the parent. When Nothing is supplied latest parent will be used.
   Maybe (StrictMaybe (GovPurposeId 'CommitteePurpose)) ->
   -- | CC members to remove
@@ -1616,7 +1646,9 @@ mkUpdateCommitteeProposal mParent ccsToRemove ccsToAdd threshold = do
   mkProposal $ UpdateCommittee parent ccsToRemove newCommitteMembers threshold
 
 submitUpdateCommittee ::
-  ConwayEraImp era =>
+  ( ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   -- | Set the parent. When Nothing is supplied latest parent will be used.
   Maybe (StrictMaybe (GovPurposeId 'CommitteePurpose)) ->
   -- | CC members to remove