+++ eras/conway/impl/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Decoupled `ConwayEraTxCert` from `ShelleyEraTxCert`, so added `ShelleyEraTxCert` constraint to:
   * `DecCBOR ConwayTxCert`
   * `transTxCert`
+  * `transTxCertV1V2`
 * Added `conwayTxCertDelegDecoder`
 * Deprecate `PoolParams` in favor of `StakePoolState`. #5196
   * Update `DRepPulser` and `RatifyEnv` to use `StakePoolState` instead of `PoolParams`.
@@ -91,6 +92,24 @@
 
 ### `testlib`
 
+* Hoisted `ShelleyEraTxCert` from `ShelleyEraImp`, so added `ShelleyEraTxCert` constraint to:
+  * `trySubmitGovActions`
+  * `mkProposal`
+  * `submitGovAction_`
+  * `submitGovActions`
+  * `submitTreasuryWithdrawals`
+  * `enactTreasuryWithdrawals`
+  * `submitParameterChange`
+  * `submitFailingGovAction`
+  * `submitCommitteeElection`
+  * `electBasicCommittee`
+  * `submitGovActionForest`
+  * `mkConstitutionProposal`
+  * `submitConstitution`
+* Decoupled `ConwayEraTxCert` from `ShelleyEraTxCert`, so added `ShelleyEraTxCert` constraint to:
+  * `genUnRegTxCert`
+  * `genRegTxCert`
+  * `setupPoolWithoutStake`
 * Added `registerPoolWithDeposit`
 * Added `registerStakeCredentialWithDeposit`
 * Added `Examples` module with `ledgerExamples`