+++ eras/shelley/impl/CHANGELOG.md
@@ -68,6 +68,11 @@
 
 ### `testlib`
 
+* Hoisted `ShelleyEraTxCert` from `ShelleyEraImp`, so added `ShelleyEraTxCert` constraint to:
+  * `registerStakeCredential`
+  * `delegateStake`
+  * `registerRewardAccount`
+  * `registerPool`
 * Added `withEachEraVersion`
 * Added `Examples` module with: `LedgerExamples`, `ledgerExamples`, `mkLedgerExamples`, `exampleCerts`,`exampleWithdrawals`, `exampleAuxDataMap`, `exampleNonMyopicRewards`, `exampleCoin`, `examplePayKey`, `exampleStakeKey`, `exampleNewEpochState`, `examplePoolDistr`, `examplePoolParams`, `exampleTxIns`, `exampleProposedPPUpdates`, `exampleByronAddress`, `testShelleyGenesis`, `keyToCredential`, `mkDSIGNKeyPair`, `mkKeyHash`, `mkScriptHash`, `mkWitnessesPreAlonzo`, `seedFromByte`, `seedFromWords`
 * Add `nativeAlwaysFails`, `nativeAlwaysSucceeds`